Split easing code into it's own module
